Moves nock setup to a function that runs before each test.
This change de-duplicates the logic for setting up the nocks for each
test.  The mock definition files are now referenced based on the title of
the test. This change makes it easier to add more tests or update
existing ones in the future.

This change also makes the tests reference use the nock definitions as
reference for the application and release IDs.

Change-type: patch
Signed-off-by: Carlo Miguel F. Cruz <carloc@balena.io>